refactor: extract areColorsSimilar() function
